+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Default and launch changed to use `about:blank` (faster start and less bandwidth) @raycardillo
+
 ### Removed
 
 ## [0.4.2] - 2025-02-11