+++ CHANGELOG.md
@@ -11,13 +11,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Removed
+
+## [0.4.3] - 2025-02-11
+
+### Added
+
 - Add logs for Chrome process output on connection failure @stephanlensky
 
 ### Changed
 
 - Default and launch changed to use `about:blank` (faster start and less bandwidth) @raycardillo
-
-### Removed
 
 ## [0.4.2] - 2025-02-11